requirements.txt
Flask
requests
beautifulsoup4
dnspython